restructure and document
