+++ client/webserver/site/src/css/market.scss
@@ -75,7 +75,7 @@ table.ordertable {
   }
 
   tbody > tr:hover {
-    background-color: #e9e9e9;
+    background-color: $light_bg_1;
   }
 
   .ico-check {
@@ -85,7 +85,7 @@ table.ordertable {
 
 .market-header {
   height: 30px;
-  background-color: #d1d1d1;
+  background-color: $light_bg_2;
   color: black;
 }
 
@@ -120,7 +120,7 @@ table.ordertable {
   padding: 0 3px;
   font-size: 14px;
   font-family: 'demi-sans', sans-serif;
-  background-color: #f9f9f9;
+  background-color: $light_bg_1;
 }
 
 #orderform {
